- Added updates from staging to live including subtitle support

+++ plugin.video.mvl/plugin.video.mvl/default.py
@@ -6,6 +6,23 @@ if 'do_nothing' in sys.argv[0]:
 
 import os
 
+#writes content to a file
+def file_write(file_name, data):
+    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), file_name)
+    f = open(file_path, 'w')
+    if data is not None:
+        f.write(data)
+    f.close()
+
+#reads content from a file
+def file_read(file_name):
+    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), file_name)
+    f = open(file_path, 'r')
+    data = f.read()
+    f.close()
+
+    return data
+
 #########
 # load last path
 file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_path.dat')
@@ -18,14 +35,12 @@ else:
 
 if last_path == sys.argv[0]:
     #same path requested again. no need to load. exit immediately
-    print "SAME PATH"
+    #print "SAME PATH"
     xbmc.executebuiltin( "Dialog.Close(busydialog)" )
     exit()
 
-file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_path.dat')
-f = open(file_path, 'w')
-f.write(sys.argv[0])
-f.close()
+file_write('screen_path.dat', sys.argv[0])
+
 ####
 
 #hide any existing loading and show system busy dialog to freeze the screen.
@@ -33,10 +48,7 @@ xbmc.executebuiltin( "Dialog.Close(busydialog)" )
 xbmc.executebuiltin( "ActivateWindow(busydialog)" )
 
 #save lockdown state to a file for future reference
-file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_lock.dat')
-f = open(file_path, 'w')
-f.write('true')
-f.close()
+file_write('screen_lock.dat', 'true')
 ####
 
 
@@ -84,7 +96,7 @@ usrsettings = xbmcaddon.Addon(id=common.plugin_id)
 # authentication['logged_in'] = 'false'
 #username = usrsettings.getSetting('username_xbmc')
 #activation_key = usrsettings.getSetting('activationkey_xbmc')
-page_limit = 30
+page_limit = 100
 username = ''
 activation_key = ''
 usrsettings.setSetting(id='mac_address', value=usrsettings.getSetting('mac_address'))
@@ -129,20 +141,16 @@ def index():
     global Main_cat
     global last_path
 
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'quit_log.dat')
-    f = open(file_path, 'w')
-    f.close()
+    file_write('quit_log.dat', None)
 
     if last_path == '':
-        file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'term_agree.dat')
-        f = open(file_path, 'w')
-        f.write('false')
-        f.close()
+        file_write('term_agree.dat', 'false')
 
     # copy pre-cached db
     src_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'resources', 'data', 'video_cache.db')
     dest_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', 'userdata', 'addon_data', 'script.module.metahandler', 'meta_cache', 'video_cache.db')
-    shutil.copyfile(src_path, dest_path)
+    if not os.path.exists(dest_path) or (os.path.exists(dest_path) and os.path.getsize(dest_path) < os.path.getsize(src_path)):
+        shutil.copyfile(src_path, dest_path)
 
     #clear Current Section name saved in the skin
     xbmc.executebuiltin('Skin.SetString(CurrentSection,)')
@@ -295,12 +303,7 @@ def setup_internet_check():
     t.start()
 
 def check_quit_log():
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'quit_log.dat')
-    f = open(file_path, 'r')
-    r = f.read()
-    f.close()
-
-    # print "check quit log = " + str(r)
+    r = file_read('quit_log.dat')
 
     if r:
         return True
@@ -380,10 +383,7 @@ def onClick_agree():
     # opener = urllib2.build_opener()
     # f = opener.open(req)
 
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'term_agree.dat')
-    f = open(file_path, 'w')
-    f.write('true')
-    f.close()
+    file_write('term_agree.dat', 'true')
 
     isAgree = True
 
@@ -409,17 +409,15 @@ def check_condition():
     # plugin.log.info(url)
     # plugin.log.info(content)
 
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'term_agree.dat')
-    f = open(file_path, 'r')
-    content = f.read()
-    f.close()
+    content = file_read('term_agree.dat')
 
     if content == 'false':
         #Show Terms & Condition window
         heading = "Terms & Conditions"
-        tc_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 't&c.info')
-        f = open(tc_path)
-        text = f.read()
+        text = file_read('t&c.info')
+
+        print 'tc-text'
+        print text
 
         #dialog = xbmcgui.Dialog()
         #agree_ret = dialog.yesno(heading, text, yeslabel='Agree', nolabel='Disagree')
@@ -430,10 +428,7 @@ def check_condition():
 
         #make sure either Agree or disagree was clicked
         #if none was clicked, then go back to home
-        file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'term_agree.dat')
-        f = open(file_path, 'r')
-        content = f.read()
-        f.close()
+        content = file_read('term_agree.dat')
         if content == 'false':
             onClick_disAgree()
 
@@ -498,10 +493,9 @@ def dialog_msg():
 def hide_busy_dialog():
     #hide loadign screen
     xbmc.executebuiltin( "Dialog.Close(busydialog)" )
+
     #clear file content to release lock
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_lock.dat')
-    f = open(file_path, 'w')
-    f.close()
+    file_write('screen_lock.dat', None)
 
 
 def show_root():
@@ -549,7 +543,7 @@ def get_categories(id, page):
             main_category_check = False
             is_search_category = False
             top_level_parent = 0
-            page_limit_cat = 30
+            page_limit_cat = 100
 
             xbmcplugin.setContent(pluginhandle, 'Movies')
 
@@ -932,7 +926,7 @@ def get_categories(id, page):
                                           'Writer': categories['writer'],
                                           'plot': mvl_plot,
                                           'genre': categories['sub_categories_names'],
-                                          'cast': categories['actors'].encode('utf-8'),
+                                          #'cast': categories['actors'].encode('utf-8'),
                                           'year': categories['release_date'],
                                           'premiered': categories['release_date'],
                                           'duration': mvl_meta['duration'],
@@ -967,7 +961,8 @@ def get_categories(id, page):
                         dp_created = True
 
                     done_count = done_count + 1
-                    dp.update((done_count*100/item_count), str(done_count)+" of "+str(item_count)+" "+dp_type+"s loaded so far.")
+                    #dp.update((done_count*100/item_count), str(done_count)+" of "+str(item_count)+" "+dp_type+"s loaded so far.")
+                    dp.update((done_count*100/item_count), str(done_count*100/item_count)+"%")
 
                     if dp.iscanceled():
                         break
@@ -1066,6 +1061,11 @@ def get_categories(id, page):
 
 @plugin.route('/get_videos/<id>/<thumbnail>/<trailer>/<parent_id>/<series_name>')
 def get_videos(id, thumbnail, trailer, parent_id, series_name):
+
+    #xbmc.executebuiltin('XBMC.RunScript(special://home\\addons\\plugin.video.mvl\\script_subtitles.py)')
+    #hide_busy_dialog()
+    #exit()
+
     if check_internet():
         show_notification()
 
@@ -1102,7 +1102,9 @@ def get_videos(id, thumbnail, trailer, parent_id, series_name):
                       }]
 
 
-            src_list = ['movreel', 'promptfile', 'mightyupload', 'novamov', 'nowvideo', 'lemupload', 'gorillavid']
+            src_list = ['movreel', 'novamov', 'nowvideo', 'gorillavid']
+            #'lemupload',
+            #'promptfile', 'mightyupload',
             #'hugefile', 'billionupload', '180upload',
             # 'firedrive', 'putlocker',
 
@@ -1193,17 +1195,6 @@ def get_videos(id, thumbnail, trailer, parent_id, series_name):
         dialog_msg()
         hide_busy_dialog()
 
-
-def getFullPath(path, url, useKiosk, userAgent):
-    profile = ""
-    # if useOwnProfile:
-    #     profile = '--user-data-dir="'+profileFolder+'" '
-    kiosk = ""
-    if useKiosk=="yes":
-        kiosk = '--kiosk '
-    if userAgent:
-        userAgent = '--user-agent="'+userAgent+'" '
-    return '"'+path+'" '+profile+userAgent+'--start-maximized --disable-translate --disable-new-tab-first-run --no-default-browser-check --no-first-run '+kiosk+'"'+url+'"'
 
 video_popup = None
 
@@ -1392,7 +1383,7 @@ def play_video(url, resolved_url, title, video_type):
 
 
             while player._playbackLock.isSet():
-                print('- - -' +'Playback lock set. Sleeping for 250.')
+                #print('- - -' +'Playback lock set. Sleeping for 250.')
                 xbmc.sleep(250)
 
             #if we are here, it means playback has either stopped or finished
@@ -1519,52 +1510,6 @@ def login_check():
 
 from xml.dom import minidom
 
-def check_update():
-    try:
-        #get latest version number from github repo
-        url = "https://raw.githubusercontent.com/ashikzk/mvl-skin/master/addons.xml"
-        req = urllib2.Request(url)
-        opener = urllib2.build_opener()
-        f = opener.open(req)
-        content = f.read()
-
-        xmldoc = minidom.parseString(content)
-        addonlist = xmldoc.getElementsByTagName('addon')
-        mvl_video_version_latest = addonlist[0].attributes['version'].value
-        mvl_skin_version_latest = addonlist[1].attributes['version'].value
-
-        #now get currently installed version numbers
-        #mvl video addon
-        file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'addon.xml')
-        xmldoc = minidom.parse(file_path)
-        addon = xmldoc.getElementsByTagName('addon')[0]
-        mvl_video_version = addon.attributes['version'].value
-
-        #mvl skin
-        file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', common.skin_id, 'addon.xml')
-        xmldoc = minidom.parse(file_path)
-        addon = xmldoc.getElementsByTagName('addon')[0]
-        mvl_skin_version = addon.attributes['version'].value
-
-        #print 'Current Version = ' + mvl_video_version + ' and ' + mvl_skin_version
-        #print 'Latest Version = ' + mvl_video_version_latest + ' and ' + mvl_skin_version_latest
-
-        if mvl_video_version == mvl_video_version_latest and mvl_skin_version == mvl_skin_version_latest:
-            #showMessage('No Update Required', 'You already have the latest version. No update is required.')
-            return False
-        else:
-            showMessage('Update Required', 'New version is available. You need to update your system before proceeding any further.')
-            return True
-
-    except IOError, e:
-        dialog_msg()
-        return False
-
-    pass
-
-def run_update():
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'script_update.py')
-    xbmc.executebuiltin("RunScript("+file_path+")")
 
 @plugin.route('/search/<category>/')
 def search(category):
@@ -1792,7 +1737,7 @@ def search(category):
                                               'Writer': categories['writer'],
                                               'plot': mvl_plot,
                                               'genre': categories['sub_categories_names'],
-                                              'cast': categories['actors'].encode('utf-8'),
+                                              #'cast': categories['actors'].encode('utf-8'),
                                               'year': categories['release_date'],
                                               'premiered': categories['release_date'],
                                               'duration': mvl_meta['duration'],
@@ -1828,7 +1773,7 @@ def search(category):
                             dp_created = True
                                   
                         done_count = done_count + 1
-                        dp.update((done_count*100/item_count), str(done_count)+" of "+str(item_count)+" "+dp_type+"s loaded so far.")
+                        dp.update((done_count*100/item_count), str(done_count*100/item_count)+"%")
 
                         if dp.iscanceled():
                             break                                 
@@ -2115,7 +2060,7 @@ def get_azlist(key, page, category):
                                           'Writer': results['writer'],
                                           'plot': mvl_plot,
                                           'genre': results['sub_categories_names'],
-                                          'cast': results['actors'].encode('utf-8'),
+                                          #'cast': results['actors'].encode('utf-8'),
                                           'year': results['release_date'],
                                           'premiered': results['release_date'],
                                           'duration': mvl_meta['duration']
@@ -2150,7 +2095,7 @@ def get_azlist(key, page, category):
                         dp_created = True
                                   
                     done_count = done_count + 1
-                    dp.update((done_count*100/item_count), str(done_count)+" of "+str(item_count)+" "+dp_type+"s loaded so far.")
+                    dp.update((done_count*100/item_count), str(done_count*100/item_count)+"%")
 
                     if dp.iscanceled():
                         break
@@ -2266,7 +2211,7 @@ def mostpopular(page, category):
                     dp_created = True
                               
                 done_count = done_count + 1
-                dp.update((done_count*100/item_count), str(done_count)+" of "+str(item_count)+" "+dp_type+"s loaded so far.")
+                dp.update((done_count*100/item_count), str(done_count*100/item_count)+"%")
 
                 if dp.iscanceled():
                     break
@@ -2320,10 +2265,7 @@ def sys_exit():
     xbmc.executebuiltin("XBMC.ActivateWindow(Home)")
 
     #reset path to home
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_path.dat')
-    f = open(file_path, 'w')
-    f.write('')
-    f.close()
+    file_write('screen_path.dat', None)
 
 
 @plugin.route('/get_favourites/<category>/')
@@ -2371,6 +2313,7 @@ def get_favourites(category):
     return items
 
 
+
 class CustomTermsPopup(xbmcgui.WindowXMLDialog):
     def __init__(self, xmlFilename, scriptPath, defaultSkin = "Default", defaultRes = "1080i"):
         pass
@@ -2389,7 +2332,6 @@ class CustomTermsPopup(xbmcgui.WindowXMLDialog):
         elif control == 10:
             self.close()
             onClick_disAgree()
-
 
 
 class CustomPopup(xbmcgui.WindowXMLDialog):
@@ -2437,6 +2379,10 @@ class CustomPopup(xbmcgui.WindowXMLDialog):
 
         elif control == 23:
             #exit
+
+            #clear path of current popup
+            file_write('screen_path.dat', None)
+
             self.close()
 
         elif control == 24:
@@ -2489,15 +2435,6 @@ class CustomPopup(xbmcgui.WindowXMLDialog):
             #self.close()
             pass
 
-            #path = 'C:\\Program Files\\Google\\Chrome\\Application\\chrome.exe'
-            #path64 = 'C:\\Program Files (x86)\\Google\\Chrome\\Application\\chrome.exe'
-            #if os.path.exists(path):
-            #    # fullUrl = getFullPath(path, "http://www.facebook.com", "", "")
-            #    # subprocess.check_call("am start -a android.intent.action.VIEW -d http://www.facebook.com", shell=False)
-            #    subprocess.check_call(path+" http://www.facebook.com", shell=False)
-            #elif os.path.exists(path64):
-            #    subprocess.check_call(path64+" http://www.facebook.com", shell=False)
-
 
 class CustomReviewPopup(xbmcgui.WindowXMLDialog):
     def __init__(self, xmlFilename, scriptPath, defaultSkin = "Default", defaultRes = "1080i"):
@@ -2520,7 +2457,7 @@ class CustomReviewPopup(xbmcgui.WindowXMLDialog):
             self.close()
             resume_popup_window()
 
-            
+
 class CustomPurchaseOptions(xbmcgui.WindowXMLDialog):
     def __init__(self, xmlFilename, scriptPath, defaultSkin = "Default", defaultRes = "1080i"):
         pass
@@ -2539,8 +2476,8 @@ class CustomPurchaseOptions(xbmcgui.WindowXMLDialog):
     def onClick	(self, control):
         if control == 10:
             self.close()
-            
-            
+
+
 class CustomKeyboard(xbmcgui.WindowXMLDialog):
     def __init__(self, xmlFilename, scriptPath, category, words, defaultSkin = "Default", defaultRes = "1080i"):
         self.isUpper  = 0
@@ -2554,7 +2491,6 @@ class CustomKeyboard(xbmcgui.WindowXMLDialog):
         self.labelString = None
         self.updateKeyboardLabel()
         self.words = words
-
 
     def showCursor(self):
         if self.isLock == 1:
@@ -2573,7 +2509,6 @@ class CustomKeyboard(xbmcgui.WindowXMLDialog):
         time.sleep(.4)
         self.showCursor()
 
-
     def updateKeyboardLabel(self):
 
         self.getControl(311).setLabel("Search Media Engine")
@@ -2588,9 +2523,6 @@ class CustomKeyboard(xbmcgui.WindowXMLDialog):
         self.t = Thread(name='test', target=self.showCursor)
         self.t.daemon = True
         self.t.start()
-
-
-
 
     def updateKeyboardLabelNumeric(self):
         self.isLock = 0;
@@ -2906,16 +2838,10 @@ if __name__ == '__main__':
     xbmc.executebuiltin("Container.SetViewMode(%s)" % mvl_view_mode)
 
     ####
-    file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_path.dat')
-    f = open(file_path, 'r')
-    last_path = f.read()
-    f.close()
+    last_path = file_read('screen_path.dat')
 
-    if 'mark_as' not in last_path:
+    if last_path and 'mark_as' not in last_path:
         #do not update path in case of mark_as_watched/unwatched was selected
         path = xbmc.getInfoLabel('Container.FolderPath')
-        file_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'screen_path.dat')
-        f = open(file_path, 'w')
-        f.write(path)
-        f.close()
+        file_write('screen_path.dat', path)
